Handle the sun directions in Solstice
Currently Solstice can only render the provided scene. The sun
directions are used to update the entities priorly to the rendering. If
several directions are submitted, Solstice renders an image per sun
direction. Each output image are then preceded by the header
"# Sun direction: ..." defining the Cartesian coordinates of the sun
direction used to update the position of the Solstice entities.