Fix the setup of anchor nodes
Entities are setup independently of the order in which they are
parsed. As a consequence, an anchor node may be not created at the setup
of a node that targets it, leading to invalid memory access.

This commit creates the anchor node when it was first referenced while
previously it was spawned on the setup of the entity that defines it.