Remove the cache of parsed entities
An parsed entity is no more cached, i.e. it is not reused. Among others
things, this ensure that the anchors are unique per entity even though
the entity is "instantiated" several times. Its id can thus be used to
map toward an unique per entity data.

This fix a bug in the solstice_setup_entities function, where an anchor
id was used to map a solstice node while this id was shared between
several entities.