Major refactoring & rewrite of the test_model project
Transform the test_model sub-project into a solstice test and rename it
in test_solstice_simulation. Replace the error managements by CHECK
directives. Provide the path toward the solstice program as an argument
of the test. In order to handle multiple build configurations, this path
is dynamically resolved by the build system.