Convert the solstice CLI man page from ronn to asciidoc
Ronn seems quite deprecated; use the asciidoc format instead. Provide a
new build process based on the asciidoc tool suite to automatically
generate the TROFF and/or the HTML documentation from the asciidoc
files. Note that the solstice-input, solstice-output and
solstice-receiver man pages are not converted in asciidoc file, yet.